prototipo-api-transcription/requirements.txt
fastapi==0.128.0
uvicorn==0.39.0
python-multipart
openai-whisper
numba
numpy
torch
tqdm
more-itertools
tiktoken